requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.9
langchain-core>=0.2.0
tavily-python>=0.3.0
firecrawl-py>=0.0.16
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
groq>=0.9.0
